+++ src/cc/View.js
@@ -1,5 +1,7 @@
 import React, { useState, useEffect, useCallback } from 'react';
 import styled from '@emotion/styled';
+import rtv from 'rtvjs';
+import { Component } from '@k8slens/extensions';
 import { useExtState } from './store/ExtStateProvider';
 import { useConfig } from './store/ConfigProvider';
 import { useAuth } from './store/AuthProvider';
@@ -8,8 +10,25 @@ import { useAddClusters } from './store/AddClustersProvider';
 import { Login } from './Login';
 import { ClusterList } from './ClusterList';
 import { AddClusters } from './AddClusters';
+import { Loader } from './Loader';
+import { ErrorPanel } from './ErrorPanel';
+import { InfoPanel } from './InfoPanel';
+import { PreferencesPanel } from './PreferencesPanel';
 import * as strings from '../strings';
 import { layout, mixinFlexColumnGaps } from './styles';
+import {
+  EXT_EVENT_CLUSTERS,
+  EXT_EVENT_KUBECONFIG,
+  extEventClustersTs,
+  extEventKubeconfigTs,
+  addExtEventHandler,
+  removeExtEventHandler,
+} from '../eventBus';
+import { normalizeUrl } from './netUtil';
+
+//
+// INTERNAL STYLED COMPONENTS
+//
 
 const Container = styled.div(function () {
   return {
@@ -21,6 +40,12 @@ const Container = styled.div(function () {
     display: 'flex',
     padding: layout.gap,
     backgroundColor: 'var(--mainBackground)',
+
+    // style all <code> elements herein
+    code: {
+      // TODO: remove once https://github.com/lensapp/lens/issues/1683 is fixed
+      fontSize: 'calc(var(--font-size) * .9)',
+    },
   };
 });
 
@@ -49,13 +74,38 @@ const MainColumn = styled.div(function () {
 const HelpColumn = styled.div(function () {
   return {
     ...getColumnStyles(),
+    justifyContent: 'space-between', // push PreferencesPanel to the bottom
     marginRight: 0,
+  };
+});
 
-    '> p': {
+const HelpContent = styled.div(function () {
+  return {
+    'h2, h3, > p': {
       marginBottom: layout.gap,
+    },
+
+    'p:last-child': {
+      marginBottom: 0,
     },
   };
 });
+
+const Title = styled.div(function () {
+  return {
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+
+    h2: {
+      marginRight: layout.gap,
+    },
+  };
+});
+
+//
+// MAIN COMPONENT
+//
 
 export const View = function () {
   //
@@ -63,7 +113,7 @@ export const View = function () {
   //
 
   const {
-    state: { baseUrl, authAccess },
+    state: { baseUrl, authAccess, addToNew, offline, savePath },
     actions: extActions,
   } = useExtState();
 
@@ -93,13 +143,25 @@ export const View = function () {
   } = useClusters();
 
   const {
-    state: { loading: addClustersLoading, error: addClustersError },
+    state: {
+      loading: addClustersLoading,
+      error: addClustersError,
+      kubeClusterAdded,
+    },
     actions: addClustersActions,
   } = useAddClusters();
 
   // @type {null|Array<Cluster>} null until clusters are loaded, then an array
   //  that represents the current selection, could be empty
   const [selectedClusters, setSelectedClusters] = useState(null);
+
+  // @type {string|null} if set, the type of extension event (from the `eventBus`) that
+  //  is currently being handled; otherwise, the extension is in its 'normal' state
+  const [activeEventType, setActiveEventType] = useState(null);
+
+  // name of the cluster added (or skipped) via single kubeConfig event; null if not
+  //  processing an EXT_EVENT_KUBECONFIG event
+  const [kubeEventClusterName, setKubeEventClusterName] = useState(null);
 
   const loading =
     configLoading ||
@@ -112,8 +174,8 @@ export const View = function () {
     clustersLoading ||
     addClustersLoading;
 
-  // in order of execution/precedence
-  const hasError =
+  // @type {string|null} in order of execution/precedence
+  const errMessage =
     configError || authError || clustersError || addClustersError || null;
 
   //
@@ -125,7 +187,7 @@ export const View = function () {
       authAccess.username = info.username;
       authAccess.password = info.password;
 
-      const url = info.baseUrl.replace(/\/$/, ''); // remove end slash if any
+      const url = normalizeUrl(info.baseUrl);
 
       authAccess.clearTokens();
       extActions.setBaseUrl(url);
@@ -160,7 +222,7 @@ export const View = function () {
   );
 
   const handleClustersAdd = useCallback(
-    function ({ savePath, offline, addToNew }) {
+    function ({ password }) {
       if (!addClustersLoading) {
         addClustersActions.addClusters({
           clusters: selectedClusters,
@@ -168,7 +230,7 @@ export const View = function () {
           baseUrl,
           config,
           username: authAccess.username,
-          password: authAccess.password,
+          password: password || authAccess.password,
           offline,
           addToNew,
         });
@@ -177,6 +239,9 @@ export const View = function () {
     [
       baseUrl,
       authAccess,
+      savePath,
+      addToNew,
+      offline,
       config,
       selectedClusters,
       addClustersLoading,
@@ -184,9 +249,80 @@ export const View = function () {
     ]
   );
 
+  const handleCloseClick = useCallback(
+    function () {
+      // reset View back to login with current auth in case still valid
+      setActiveEventType(null);
+      setKubeEventClusterName(null);
+      configActions.reset();
+      authActions.reset();
+      clustersActions.reset();
+      setSelectedClusters([]);
+      addClustersActions.reset();
+    },
+    [configActions, authActions, clustersActions, addClustersActions]
+  );
+
+  const handleClustersEvent = useCallback(
+    function (event) {
+      console.log('[View] received clusters event', event); // DEBUG
+      rtv.verify({ event }, { event: extEventClustersTs });
+
+      const { data } = event;
+
+      authAccess.username = data.username;
+      authAccess.password = null;
+      authAccess.updateTokens(data.tokens);
+
+      const url = normalizeUrl(data.baseUrl);
+      extActions.setBaseUrl(url);
+      extActions.setAuthAccess(authAccess);
+
+      authActions.reset();
+      clustersActions.reset();
+
+      setActiveEventType(EXT_EVENT_CLUSTERS);
+      configActions.load(url); // implicit reset of current config
+    },
+    [authAccess, authActions, extActions, clustersActions, configActions]
+  );
+
+  const handleKubeconfigEvent = useCallback(
+    function (event) {
+      console.log('[View] received kubeconfig event', event); // DEBUG
+      rtv.verify({ event }, { event: extEventKubeconfigTs });
+
+      const { data } = event;
+
+      if (!addClustersLoading) {
+        setActiveEventType(EXT_EVENT_KUBECONFIG);
+        setKubeEventClusterName(`${data.namespace}/${data.clusterName}`);
+        addClustersActions.addKubeCluster({
+          savePath,
+          addToNew,
+          ...data,
+        });
+      }
+    },
+    [savePath, addToNew, addClustersLoading, addClustersActions]
+  );
+
   //
   // EFFECTS
   //
+
+  useEffect(
+    function () {
+      addExtEventHandler(EXT_EVENT_CLUSTERS, handleClustersEvent);
+      addExtEventHandler(EXT_EVENT_KUBECONFIG, handleKubeconfigEvent);
+
+      return function () {
+        removeExtEventHandler(EXT_EVENT_CLUSTERS, handleClustersEvent);
+        removeExtEventHandler(EXT_EVENT_KUBECONFIG, handleKubeconfigEvent);
+      };
+    },
+    [handleClustersEvent, handleKubeconfigEvent]
+  );
 
   // 1. load the config object
   useEffect(
@@ -213,7 +349,7 @@ export const View = function () {
         !authLoading &&
         !authLoaded
       ) {
-        if (authAccess.isValid()) {
+        if (authAccess.isValid(!activeEventType)) {
           // skip authentication, go straight for the clusters
           authActions.setAuthenticated();
         } else if (authAccess.hasCredentials()) {
@@ -235,6 +371,7 @@ export const View = function () {
       authAccess,
       baseUrl,
       config,
+      activeEventType,
     ]
   );
 
@@ -247,7 +384,7 @@ export const View = function () {
         baseUrl &&
         config &&
         authLoaded &&
-        authAccess.isValid()
+        authAccess.isValid(!activeEventType)
       ) {
         clustersActions.load(baseUrl, config, authAccess);
       } else if (authAccess.changed) {
@@ -263,6 +400,7 @@ export const View = function () {
       clustersLoading,
       clustersLoaded,
       clustersActions,
+      activeEventType,
     ]
   );
 
@@ -283,20 +421,80 @@ export const View = function () {
   // RENDER
   //
 
+  // DEBUG TODO: don't show Login if responding to a URL request...
+
+  const title =
+    activeEventType === EXT_EVENT_KUBECONFIG
+      ? strings.view.main.titles.kubeConfig()
+      : strings.view.main.titles.generic();
+
+  const loaderMessage = activeEventType
+    ? activeEventType === EXT_EVENT_CLUSTERS
+      ? strings.view.main.loaders.clustersHtml()
+      : strings.view.main.loaders.kubeConfig()
+    : undefined;
+
   return (
     <Container className="lecc-View">
       <MainColumn>
-        <h2>{strings.view.main.title()}</h2>
-        <Login
-          loading={loading && !addClustersLoading}
-          disabled={loading}
-          baseUrl={baseUrl || undefined}
-          username={authAccess ? authAccess.username : undefined}
-          password={authAccess ? authAccess.password : undefined}
-          onLogin={handleLogin}
-        />
-        {!hasError &&
-        authAccess.isValid() &&
+        {/* include X (close) only if we're handling an extension event */}
+        <Title>
+          <h2>{title}</h2>
+          {activeEventType && (
+            <Component.Icon
+              material="close"
+              interactive
+              big
+              title={strings.view.main.close}
+              onClick={handleCloseClick}
+            />
+          )}
+        </Title>
+
+        {/* show loader only if we ARE handling an extension event */}
+        {loading && activeEventType && <Loader message={loaderMessage} />}
+
+        {/* only show Login if we are NOT handling an extension event */}
+        {!activeEventType && (
+          <Login
+            loading={loading && !addClustersLoading}
+            disabled={loading}
+            baseUrl={baseUrl || undefined}
+            username={authAccess ? authAccess.username : undefined}
+            password={authAccess ? authAccess.password : undefined}
+            onLogin={handleLogin}
+          />
+        )}
+
+        {/* show error in UI on top of notification (since they disappear) when handling an event */}
+        {activeEventType && !!errMessage && (
+          <ErrorPanel>{errMessage}</ErrorPanel>
+        )}
+
+        {/* when handling the kubeConfig event, show the result in the UI */}
+        {activeEventType === EXT_EVENT_KUBECONFIG &&
+          !errMessage &&
+          kubeClusterAdded && (
+            <InfoPanel>
+              {strings.view.main.kubeConfigEvent.clusterAdded(
+                kubeEventClusterName
+              )}
+            </InfoPanel>
+          )}
+        {activeEventType === EXT_EVENT_KUBECONFIG &&
+          !errMessage &&
+          !kubeClusterAdded && (
+            <InfoPanel>
+              {strings.view.main.kubeConfigEvent.clusterSkipped(
+                kubeEventClusterName
+              )}
+            </InfoPanel>
+          )}
+
+        {/* ClusterList and AddClusters apply only if NOT loading a kubeConfig */}
+        {activeEventType !== EXT_EVENT_KUBECONFIG &&
+        !errMessage &&
+        authAccess.isValid(!activeEventType) &&
         clustersLoaded &&
         selectedClusters ? (
           <>
@@ -308,14 +506,19 @@ export const View = function () {
             />
             <AddClusters
               clusters={selectedClusters}
+              passwordRequired={activeEventType === EXT_EVENT_CLUSTERS}
               onAdd={handleClustersAdd}
             />
           </>
         ) : undefined}
       </MainColumn>
-      <HelpColumn
-        dangerouslySetInnerHTML={{ __html: strings.view.help.html() }}
-      />
+
+      <HelpColumn>
+        <HelpContent
+          dangerouslySetInnerHTML={{ __html: strings.view.help.html() }}
+        />
+        <PreferencesPanel />
+      </HelpColumn>
     </Container>
   );
 };