+++ src/renderer.tsx
@@ -4,6 +4,12 @@ import { LensRendererExtension } from '@k8slens/extensions';
 import { ContainerCloudIcon, AddClusterPage } from './page';
 import * as strings from './strings';
 import { addRoute } from './routes';
+import {
+  EXT_EVENT_CLUSTERS,
+  EXT_EVENT_KUBECONFIG,
+  dispatchExtEvent,
+} from './eventBus';
+import pkg from '../package.json';
 
 const itemColor = 'white'; // CSS color; Lens hard-codes the color of the workspace indicator item to 'white' also
 
@@ -28,15 +34,102 @@ export default class ExtensionRenderer extends LensRendererExtension {
 
   statusBarItems = [
     {
+      // DEBUG REMOVE once have protocol handler...
+      item: (
+        <Item
+          className="flex align-center gaps"
+          onClick={() => {
+            this.navigate(addRoute);
+
+            // DEBUG TODO test this again
+            dispatchExtEvent({
+              type: EXT_EVENT_CLUSTERS,
+              data: {
+                username: 'writer',
+                baseUrl: 'http://foo.com',
+                tokens: {
+                  id_token: 'asdf',
+                  expires_in: Date.now(),
+                  refresh_token: 'zxcv',
+                  refresh_expires_in: Date.now(),
+                },
+              },
+            });
+          }}
+        >
+          <span>EVENT</span>
+        </Item>
+      ),
+    },
+    {
       item: (
         <Item
           className="flex align-center gaps"
           onClick={() => this.navigate(addRoute)}
         >
-          <ContainerCloudIcon fill={itemColor} />
-          <span>{strings.extension.statusBar['label']()}</span>
+          <ContainerCloudIcon
+            fill={itemColor}
+            title={strings.extension.statusBar['label']()}
+          />
         </Item>
       ),
     },
   ];
+
+  onProtocolClusters = ({ search }) => {
+    let tokens;
+    try {
+      tokens = JSON.parse(atob(search.tokens));
+    } catch (err) {
+      // eslint-disable-next-line no-console -- log error
+      console.error(
+        `[${pkg.name}/renderer/onProtocolClusters] ERROR: Failed to decode tokens, error="${err.message}"`,
+        err
+      );
+      return;
+    }
+
+    this.navigate(addRoute);
+
+    dispatchExtEvent({
+      type: EXT_EVENT_CLUSTERS,
+      data: {
+        username: search.username,
+        baseUrl: search.baseUrl,
+        tokens,
+      },
+    });
+  };
+
+  onProtocolKubeConfig = ({ search }) => {
+    let kubeConfig;
+    try {
+      kubeConfig = JSON.parse(atob(search.kubeConfig));
+    } catch (err) {
+      // eslint-disable-next-line no-console -- log error
+      console.error(
+        `[${pkg.name}/renderer/onProtocolKubeConfig] ERROR: Failed to decode kubeConfig, error="${err.message}"`,
+        err
+      );
+      return;
+    }
+
+    this.navigate(addRoute);
+
+    dispatchExtEvent({
+      type: EXT_EVENT_KUBECONFIG,
+      data: {
+        namespace: search.namespace,
+        clusterName: search.clusterName,
+        clusterId: search.clusterId,
+        kubeConfig,
+      },
+    });
+  };
+
+  onActivate() {
+    // DEBUG TODO: TSC complaining it doesn't know about `onProtocolRequest` on the base class...
+    // this.onProtocolRequest(`/${EXT_EVENT_CLUSTERS}`, this.onProtocolClusters);
+    // this.onProtocolRequest(`/${EXT_EVENT_KUBECONFIG}`, this.onProtocolKubeConfig);
+  }
 }